Fully split :classic and :modern implementations
